#3766 uningest published file
