+++ src/main/java/edu/harvard/iq/dataverse/api/Files.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import edu.harvard.iq.dataverse.DataFile;
 import edu.harvard.iq.dataverse.DataFileServiceBean;
+import edu.harvard.iq.dataverse.Dataset;
 import edu.harvard.iq.dataverse.DatasetServiceBean;
 import edu.harvard.iq.dataverse.DatasetVersionServiceBean;
 import edu.harvard.iq.dataverse.DataverseRequestServiceBean;
@@ -22,7 +23,10 @@ import edu.harvard.iq.dataverse.engine.command.impl.DeleteMapLayerMetadataComman
 import edu.harvard.iq.dataverse.engine.command.impl.RestrictFileCommand;
 import edu.harvard.iq.dataverse.engine.command.impl.UningestFileCommand;
 import edu.harvard.iq.dataverse.engine.command.impl.UpdateDatasetCommand;
+import edu.harvard.iq.dataverse.export.ExportException;
+import edu.harvard.iq.dataverse.export.ExportService;
 import edu.harvard.iq.dataverse.ingest.IngestServiceBean;
+import edu.harvard.iq.dataverse.settings.SettingsServiceBean;
 import edu.harvard.iq.dataverse.util.SystemConfig;
 import java.io.InputStream;
 import java.util.ResourceBundle;
@@ -71,6 +75,8 @@ public class Files extends AbstractApiBean {
     UserNotificationServiceBean userNotificationService;
     @EJB
     SystemConfig systemConfig;
+    @EJB
+    SettingsServiceBean settingsService;
     
     private static final Logger logger = Logger.getLogger(Files.class.getName());
     
@@ -308,12 +314,33 @@ public class Files extends AbstractApiBean {
         try {
             DataverseRequest req = createDataverseRequest(findUserOrDie());
             execCommand(new UningestFileCommand(req, dataFile));
+            dataFile = fileService.find(idSupplied);
+            Dataset theDataset = dataFile.getOwner();
+            exportMetadata(settingsService, theDataset);
             return ok("Datafile " + idSupplied + " uningested.");
-
         } catch (WrappedResponse wr) {
             return wr.getResponse();
         }
 
+    }
+            
+    /**
+     * Attempting to run metadata export, for all the formats for which we have
+     * metadata Exporters.
+     */
+    private void exportMetadata(SettingsServiceBean settingsServiceBean, Dataset theDataset) {
+
+        try {
+            ExportService instance = ExportService.getInstance(settingsServiceBean);
+            instance.exportAllFormats(theDataset);
+
+        } catch (ExportException ex) {
+            // Something went wrong!
+            // Just like with indexing, a failure to export is not a fatal
+            // condition. We'll just log the error as a warning and keep
+            // going:
+            logger.log(Level.WARNING, "Dataset publication finalization: exception while exporting:{0}", ex.getMessage());
+        }
     }
 
 }
